fixed regression from #832, where xlsx files where not correctly identified, using the wrong fallback.
Updated tests, and migrated last remaining tests from gnrlist_test to flatfiles, in the
correct context.

Legacy tests are still in place since we still support read* functions, although
they're deprecated

refs #1002
