+++ gnrpy/tests/core/gnrlist_test.py
@@ -1,7 +1,6 @@
 import os.path
 import datetime
 import tempfile
-import csv
     
 import pytest
 from gnr.core import gnrlist as gl
@@ -157,159 +156,6 @@ def test_sortByItem():
     res = gl.sortByItem(test_l, "company.name:d*", hkeys=True)
     assert "Wayne" in res[0]['company']['name']
     
-def test_getReader():
-
-    with tempfile.TemporaryDirectory() as tmpdir:
-        filename = os.path.join(tmpdir, 'test.csv')
-        with open(filename, "w") as wfp:
-            wfp.write("one,two,three\nfour,five,six")
-        a = gl.getReader(filename)
-        assert isinstance(a, gl.CsvReader)
-        a.filecsv.close()
-
-        filename = os.path.join(tmpdir, 'test.tab')
-        with open(filename, "w") as wfp:
-            wfp.write("one\ttwo\tthree\nfour\tfive\tsix")
-        a = gl.getReader(filename)
-        assert isinstance(a, gl.CsvReader)
-        a.filecsv.close()
-        
-        filename = os.path.join(tmpdir, 'test.csv')
-        with open(filename, "w") as wfp:
-            wfp.write("one\ttwo\tthree\nfour\tfive\tsix")
-        a = gl.getReader(filename, filetype="csv_auto")
-        assert isinstance(a, gl.CsvReader)
-        a.filecsv.close()
-
-        # Will fail with an emtpy file
-        filename = os.path.join(tmpdir, 'test.xls')
-        with open(filename, "w") as wfp:
-            pass
-        with pytest.raises(Exception):
-            a = gl.getReader(filename, filetype="excel")
-            a.filecsv.close()
-        filename = os.path.join(tmpdir, 'test.xlsx')
-        with open(filename, "w") as wfp:
-            pass
-        with pytest.raises(Exception):
-            a = gl.getReader(filename, filetype="excel")
-            a.filecsv.close()
-            
-    test_dir = os.path.dirname(__file__)
-    
-    filename = os.path.join(test_dir, "data", "test.xls")
-    a = gl.getReader(filename)
-    assert isinstance(a, gl.XlsReader)
-
-    filename = os.path.join(test_dir, "data","test.xlsx")
-    a = gl.getReader(filename)
-    assert isinstance(a, gl.XlsxReader)
-
-    # FIXME: this fails all the time.
-    with pytest.raises(Exception):
-        filename = os.path.join(test_dir, "data", "testbag.xml")
-        a = gl.getReader(filename)
-
-def test_CsvReader():
-    test_dir = os.path.dirname(__file__)
-    test_file = os.path.join(test_dir, "data", "test.csv")
-    a = gl.CsvReader(test_file)
-    # FIXME: odd interface using __call__
-    r = [x for x in a()]
-    assert len(r) == 1
-    assert isinstance(r[0], gl.GnrNamedList)
-    assert 'a' in r[0].keys()
-    a = gl.CsvReader(test_file, detect_encoding=True)
-
-
-def test_XlsReader():
-    test_dir = os.path.dirname(__file__)
-    test_file = os.path.join(test_dir, "data", 'test.xls')
-    r = gl.XlsReader(test_file)
-    assert r.sheet.name == "Sheet1"
-    assert 'a' in r.headers
-    assert 0 in r.colindex
-    assert r.colindex[0] is True
-    assert 'a' in r.index
-    assert r.ncols == 3
-    assert r.nrows == 1
-    d = [x for x in r()]
-    assert len(d) == 1
-    assert isinstance(d[0], gl.GnrNamedList)
-    assert 'a' in d[0].keys()
-
-def test_XlsxReader():
-    test_dir = os.path.dirname(__file__)
-    test_file = os.path.join(test_dir, "data", 'test.xlsx')
-    r = gl.XlsxReader(test_file)
-    assert r.sheet.title == "Sheet1"
-    assert 'a' in r.headers
-    assert 0 in r.colindex
-    assert r.colindex[0] is True
-    assert 'a' in r.index
-    assert r.ncols == 3
-
-    # FIXME: this doesn't work in the implementation
-    #assert r.nrows == 1
-
-    d = [x for x in r()]
-    assert len(d) == 1
-    assert isinstance(d[0], gl.GnrNamedList)
-    assert 'a' in d[0].keys()
-
-def test_readXLS():
-    test_dir = os.path.dirname(__file__)
-    test_file = os.path.join(test_dir, "data", 'test.xls')
-    r = gl.readXLS(test_file)
-    d = [x for x in r]
-    assert len(d) == 1
-    assert isinstance(d[0], gl.GnrNamedList)
-    assert 'a' in d[0].keys()
-
-    with open(test_file, "rb") as fp:
-        r = gl.readXLS(fp)
-        d = [x for x in r]
-        assert len(d) == 1
-        assert isinstance(d[0], gl.GnrNamedList)
-        assert 'a' in d[0].keys()
-
-def test_readCSV():
-    # FIXME: apparently, readXLS and readCSV exposes
-    # a different interface to access record, please
-    # check the last assert here with the last of readXLS test
-    # while readCSV_new works correctly. Maybe the _new should
-    # be the implementation..
-    test_dir = os.path.dirname(__file__)
-    test_file = os.path.join(test_dir, "data", 'test.csv')
-    r = gl.readCSV(test_file)
-    d = [x for x in r]
-    assert len(d) == 2
-    assert isinstance(d[0], gl.GnrNamedList)
-    assert 'a' in d[0].keys()[0]
-
-    with open(test_file, "r") as fp:
-        r = gl.readCSV(fp)
-        d = [x for x in r]
-        assert len(d) == 2
-        assert isinstance(d[0], gl.GnrNamedList)
-        assert 'a' in d[0].keys()[0]
-
-def test_readCSV_new():
-    test_dir = os.path.dirname(__file__)
-    test_file = os.path.join(test_dir, "data", 'test.csv')
-    r = gl.readCSV_new(test_file)
-    d = [x for x in r]
-    assert len(d) == 1
-    assert isinstance(d[0], gl.GnrNamedList)
-    assert 'a' in d[0].keys()
-
-    with open(test_file, "r") as fp:
-        r = gl.readCSV_new(fp)
-        d = [x for x in r]
-        assert len(d) == 1
-        assert isinstance(d[0], gl.GnrNamedList)
-        assert 'a' in d[0].keys()
-
         
 def test_sortByAttr():
     class MockObj(object):
@@ -329,168 +175,6 @@ def test_sortByAttr():
     m2.a = MockObj()
 
 
-def test_CsvReader_duplicate_columns():
-    """Test handling of duplicate column names in CSV files"""
-
-    with tempfile.NamedTemporaryFile(mode='w', suffix='.csv', delete=False, newline='') as f:
-        csv_file = f.name
-        writer = csv.writer(f)
-        # Headers with duplicate 'name' column
-        writer.writerow(['id', 'name', 'surname', 'name', 'email'])
-        writer.writerow(['1', 'Mario', 'Rossi', 'Giuseppe', 'mario@test.com'])
-        writer.writerow(['2', 'Laura', 'Bianchi', 'Anna', 'laura@test.com'])
-
-    try:
-        reader = gl.CsvReader(csv_file)
-
-        # Check that duplicate column has been renamed
-        assert 'name' in reader.headers
-        assert 'name[3]' in reader.headers
-        assert reader.headers == ['id', 'name', 'surname', 'name[3]', 'email']
-
-        # Check index mapping
-        assert reader.index['name'] == 1
-        assert reader.index['name[3]'] == 3
-
-        # Read rows
-        rows = [row for row in reader()]
-        assert len(rows) == 2
-
-        # Test first row
-        row = rows[0]
-        assert row[0] == '1'
-        assert row[1] == 'Mario'
-        assert row[2] == 'Rossi'
-        assert row[3] == 'Giuseppe'
-        assert row[4] == 'mario@test.com'
-
-        # Access by name
-        assert row['id'] == '1'
-        assert row['name'] == 'Mario'
-        assert row['surname'] == 'Rossi'
-        assert row['name[3]'] == 'Giuseppe'
-        assert row['email'] == 'mario@test.com'
-
-    finally:
-        os.unlink(csv_file)
-
-
-def test_XlsxReader_duplicate_columns():
-    """Test handling of duplicate column names in XLSX files"""
-    try:
-        from openpyxl import Workbook
-    except ImportError:
-        pytest.skip("openpyxl not available")
-
-    with tempfile.NamedTemporaryFile(suffix='.xlsx', delete=False) as f:
-        xlsx_file = f.name
-
-    try:
-        # Create XLSX with duplicate columns
-        wb = Workbook()
-        ws = wb.active
-        ws.append(['id', 'name', 'surname', 'name', 'email'])
-        ws.append(['1', 'Mario', 'Rossi', 'Giuseppe', 'mario@test.com'])
-        ws.append(['2', 'Laura', 'Bianchi', 'Anna', 'laura@test.com'])
-        wb.save(xlsx_file)
-
-        reader = gl.XlsxReader(xlsx_file)
-
-        # Check that duplicate column has been renamed
-        assert 'name' in reader.headers
-        assert 'name[3]' in reader.headers
-        assert reader.headers == ['id', 'name', 'surname', 'name[3]', 'email']
-
-        # Check index mapping
-        assert reader.index['name'] == 1
-        assert reader.index['name[3]'] == 3
-
-        # Read rows
-        rows = [row for row in reader()]
-        assert len(rows) == 2
-
-        # Test first row
-        row = rows[0]
-        assert row[0] == '1'
-        assert row[1] == 'Mario'
-        assert row[2] == 'Rossi'
-        assert row[3] == 'Giuseppe'
-        assert row[4] == 'mario@test.com'
-
-        # Access by name
-        assert row['id'] == '1'
-        assert row['name'] == 'Mario'
-        assert row['surname'] == 'Rossi'
-        assert row['name[3]'] == 'Giuseppe'
-        assert row['email'] == 'mario@test.com'
-
-    finally:
-        os.unlink(xlsx_file)
-
-
-def test_XlsReader_duplicate_columns():
-    """Test handling of duplicate column names in XLS files"""
-    try:
-        import xlwt
-    except ImportError:
-        pytest.skip("xlwt not available")
-
-    with tempfile.NamedTemporaryFile(suffix='.xls', delete=False) as f:
-        xls_file = f.name
-
-    try:
-        # Create XLS with duplicate columns
-        wb = xlwt.Workbook()
-        ws = wb.add_sheet('Sheet1')
-
-        # Headers
-        headers = ['id', 'name', 'surname', 'name', 'email']
-        for col, header in enumerate(headers):
-            ws.write(0, col, header)
-
-        # Data rows
-        data = [
-            ['1', 'Mario', 'Rossi', 'Giuseppe', 'mario@test.com'],
-            ['2', 'Laura', 'Bianchi', 'Anna', 'laura@test.com']
-        ]
-        for row_idx, row_data in enumerate(data, start=1):
-            for col_idx, value in enumerate(row_data):
-                ws.write(row_idx, col_idx, value)
-
-        wb.save(xls_file)
-
-        reader = gl.XlsReader(xls_file)
-
-        # Check that duplicate column has been renamed
-        assert 'name' in reader.headers
-        assert 'name[3]' in reader.headers
-        assert reader.headers == ['id', 'name', 'surname', 'name[3]', 'email']
-
-        # Check index mapping
-        assert reader.index['name'] == 1
-        assert reader.index['name[3]'] == 3
-
-        # Read rows
-        rows = [row for row in reader()]
-        assert len(rows) == 2
-
-        # Test first row
-        row = rows[0]
-        assert row[0] == '1'
-        assert row[1] == 'Mario'
-        assert row[2] == 'Rossi'
-        assert row[3] == 'Giuseppe'
-        assert row[4] == 'mario@test.com'
-
-        # Access by name
-        assert row['id'] == '1'
-        assert row['name'] == 'Mario'
-        assert row['surname'] == 'Rossi'
-        assert row['name[3]'] == 'Giuseppe'
-        assert row['email'] == 'mario@test.com'
-
-    finally:
-        os.unlink(xls_file)
 
 
 def test_hGetAttr():
@@ -524,25 +208,6 @@ def test_hGetAttr():
     assert gl.hGetAttr(obj, 'profile.city') is None
 
 
-def test_readTab():
-    """Test tab-delimited file reading"""
-    with tempfile.NamedTemporaryFile(mode='w', suffix='.tab', delete=False) as f:
-        tab_file = f.name
-        f.write("name\tage\tcity\n")
-        f.write("Alice\t30\tNYC\n")
-        f.write("Bob\t25\tLA\n")
-
-    try:
-        rows = list(gl.readTab(tab_file))
-        assert len(rows) == 2
-        assert isinstance(rows[0], gl.GnrNamedList)
-        assert rows[0]['name'] == 'Alice'
-        assert rows[0]['age'] == '30'
-        assert rows[0]['city'] == 'NYC'
-        assert rows[1]['name'] == 'Bob'
-
-    finally:
-        os.unlink(tab_file)
 
 
 def test_GnrNamedList_extractMethods():
@@ -590,35 +255,6 @@ def test_GnrNamedList_dynamic_columns():
     assert len(row._index) == 4
 
 
-def test_multiple_duplicate_columns():
-    """Test handling of 3+ duplicate columns"""
-
-    with tempfile.NamedTemporaryFile(mode='w', suffix='.csv', delete=False, newline='') as f:
-        csv_file = f.name
-        writer = csv.writer(f)
-        # Three 'name' columns
-        writer.writerow(['id', 'name', 'name', 'name', 'email'])
-        writer.writerow(['1', 'First', 'Middle', 'Last', 'test@test.com'])
-
-    try:
-        reader = gl.CsvReader(csv_file)
-
-        # Check all duplicate columns are renamed
-        assert reader.headers == ['id', 'name', 'name[2]', 'name[3]', 'email']
-        assert reader.index['name'] == 1
-        assert reader.index['name[2]'] == 2
-        assert reader.index['name[3]'] == 3
-
-        rows = list(reader())
-        row = rows[0]
-
-        # All values accessible
-        assert row['name'] == 'First'
-        assert row['name[2]'] == 'Middle'
-        assert row['name[3]'] == 'Last'
-
-    finally:
-        os.unlink(csv_file)
 
 
 def test_GnrNamedList_iteritems():
@@ -659,91 +295,6 @@ def test_sortByItem_case_insensitive():
     assert result[1]['name'] == 'bob'
     assert result[2]['name'] == 'Charlie'
 
-
-def test_slugify_consistency_across_readers():
-    """Test that CSV, XLS, and XLSX readers use consistent slugification"""
-
-    # Headers with spaces and special characters
-    headers = ['Transaction ID', 'User Name', 'Email-Address', 'Created At']
-    expected_keys = ['transaction_id', 'user_name', 'email_address', 'created_at']
-
-    # Test CSV via getReader
-    with tempfile.NamedTemporaryFile(mode='w', suffix='.csv', delete=False, newline='') as f:
-        csv_file = f.name
-        writer = csv.writer(f)
-        writer.writerow(headers)
-        writer.writerow(['123', 'Alice', 'alice@test.com', '2025-10-28'])
-
-    try:
-        csv_reader = gl.getReader(csv_file)
-        csv_keys = list(csv_reader.index.keys())
-        assert csv_keys == expected_keys, f"CSV keys: {csv_keys}"
-
-        # Verify access works with underscores
-        for row in csv_reader():
-            assert row['transaction_id'] == '123'
-            assert row['user_name'] == 'Alice'
-            break
-    finally:
-        os.unlink(csv_file)
-
-    # Test XLSX
-    try:
-        from openpyxl import Workbook
-    except ImportError:
-        pytest.skip("openpyxl not available")
-
-    with tempfile.NamedTemporaryFile(suffix='.xlsx', delete=False) as f:
-        xlsx_file = f.name
-
-    try:
-        wb = Workbook()
-        ws = wb.active
-        ws.append(headers)
-        ws.append(['123', 'Alice', 'alice@test.com', '2025-10-28'])
-        wb.save(xlsx_file)
-
-        xlsx_reader = gl.XlsxReader(xlsx_file)
-        xlsx_keys = list(xlsx_reader.index.keys())
-        assert xlsx_keys == expected_keys, f"XLSX keys: {xlsx_keys}"
-
-        # Verify access works with underscores
-        for row in xlsx_reader():
-            assert row['transaction_id'] == '123'
-            assert row['user_name'] == 'Alice'
-            break
-    finally:
-        os.unlink(xlsx_file)
-
-    # Test XLS
-    try:
-        import xlwt
-    except ImportError:
-        pytest.skip("xlwt not available")
-
-    with tempfile.NamedTemporaryFile(suffix='.xls', delete=False) as f:
-        xls_file = f.name
-
-    try:
-        wb = xlwt.Workbook()
-        ws = wb.add_sheet('Sheet1')
-        for col, header in enumerate(headers):
-            ws.write(0, col, header)
-        for col, value in enumerate(['123', 'Alice', 'alice@test.com', '2025-10-28']):
-            ws.write(1, col, value)
-        wb.save(xls_file)
-
-        xls_reader = gl.XlsReader(xls_file)
-        xls_keys = list(xls_reader.index.keys())
-        assert xls_keys == expected_keys, f"XLS keys: {xls_keys}"
-
-        # Verify access works with underscores
-        for row in xls_reader():
-            assert row['transaction_id'] == '123'
-            assert row['user_name'] == 'Alice'
-            break
-    finally:
-        os.unlink(xls_file)
 
 
 def test_GnrNamedList_sql_adapter_compatibility():
